Complete the entire project except for finishing touches
At this point, the only things that are not quire right are as follows:
1. As of now, when the window is wide enough to house 2 rows of 3 inside
the projects-container grid, the main grid does not take up the entire page.
I will try to fix this by creating a flex container around the main-grid
and setting flex-grow to 1.
2. The icons on the left of the screen (sidebar) are supposed to be white in color
with no background color specified. I didn't know how to obtain these, so I went with a light
background color. It isn't perfect, but it is looking ok.