Fix sizing issue for main grid
The main grid was not filling up the viewport when wide enough
to house the cards without extending that far vertically. However,
we wanted the sidebar to extend to the bottom of the viewport nomatter what,
and for the entire grid to grow beyond the length of the viewport if needed
(in the case where the window was narrow enough to cause the cards to overflow).
I achieved this by setting body height to 100%, and then the main-grid's height to
100% as well, but with a min-height set to 100vh. This ensured that the main-grid
always at least filled the view port but was able to grow dynamically.